requirements.txt
crewai==0.76.2
python-dotenv==1.0.1
streamlit==1.33.0
openai==1.52.2
anthropic